extract NewConversation to new file
